requirements.txt
requests>=2.32.0
pandas>=2.3.0
pyarrow>=21.0.0
google-cloud-bigquery>=3.38.0
pandas-gbq>=0.29.0
db-dtypes>=1.4.0
sqlalchemy>=2.0.0
python-dateutil>=2.9.0
python-dotenv>=1.1.0
numpy>=2.3.0
scikit-learn>=1.5.0       